Refactor lister implementation
