+++ src/object-engine/filestore/src/fs/store.rs
@@ -16,7 +16,7 @@ use std::path::PathBuf;
 
 use tokio::fs;
 
-use super::Tenant;
+use super::{DirLister, Tenant};
 use crate::{async_trait, Error, Result};
 
 pub struct Store {
@@ -39,7 +39,7 @@ impl crate::Store for Store {
 
     async fn list_tenants(&self) -> Result<Box<dyn crate::Lister<Item = String>>> {
         let dir = fs::read_dir(&self.path).await?;
-        Ok(Box::new(Lister { dir }))
+        Ok(Box::new(DirLister::new(dir)))
     }
 
     async fn create_tenant(&self, name: &str) -> Result<Box<dyn crate::Tenant>> {
@@ -55,30 +55,5 @@ impl crate::Store for Store {
         let path = self.path.join(name);
         fs::remove_dir_all(&path).await?;
         Ok(())
-    }
-}
-
-struct Lister {
-    dir: fs::ReadDir,
-}
-
-#[async_trait]
-impl crate::Lister for Lister {
-    type Item = String;
-
-    async fn next(&mut self, n: usize) -> Result<Vec<Self::Item>> {
-        let mut result = Vec::new();
-        for _i in 0..n {
-            if let Some(ent) = self.dir.next_entry().await? {
-                let file_name = ent
-                    .file_name()
-                    .into_string()
-                    .map_err(|s| Error::Corrupted(format!("invalid name {:?}", s)))?;
-                result.push(file_name);
-            } else {
-                break;
-            }
-        }
-        Ok(result)
     }
 }
